Got rid of unused files, disabled psot button if user enters nothing in the textbox, general code clean up and made it so that the annotation appears as a first level comment box
